comments issues
